Moved compressor tests in seperate files
